introduction of plots-class
