+++ rt1/rtfits.py
@@ -24,7 +24,7 @@ from .scatter import Scatter
 from .rt1 import RT1
 
 from functools import partial, update_wrapper
-from .rtplots import printsig0timeseries, printresults, printerr
+from .rtplots import plots
 
 import copy  # used to copy objects
 import datetime
@@ -122,13 +122,14 @@ class Fits(Scatter):
         self.result = None
 
         # add plotfunctions
-        self.printsig0timeseries = partial(printsig0timeseries, fit = self)
-        update_wrapper(self.printsig0timeseries, printsig0timeseries)
-        self.printresults = partial(printresults, fit = self)
-        update_wrapper(self.printresults, printresults)
-        self.printerr = partial(printerr, fit = self)
-        update_wrapper(self.printerr, printerr)
+        # self.printsig0timeseries = partial(printsig0timeseries, fit = self)
+        # update_wrapper(self.printsig0timeseries, printsig0timeseries)
+        # self.printresults = partial(printresults, fit = self)
+        # update_wrapper(self.printresults, printresults)
+        # self.printerr = partial(printerr, fit = self)
+        # update_wrapper(self.printerr, printerr)
 
+        self.plots = plots(self)
 
     def generate_dyn_dict(self, param_keys, datetime_index,
                           freq=None, freqkeys=[],
@@ -988,15 +989,10 @@ class Fits(Scatter):
         '''
         # set up the dictionary for storing intermediate results
         if intermediate_results is True:
-            from .rtplots import plot_interres
-            self.plot_interres = partial(plot_interres, fit = self)
-            update_wrapper(self.plot_interres, plot_interres)
-
             if not hasattr(self, 'intermediate_results'):
                 self.intermediate_results = {'parameters':[],
                                              'residuals':[],
                                              'jacobian':[]}
-
 
 
         # generate a list of the names of the parameters that will be fitted.
@@ -1564,570 +1560,3 @@ class Fits(Scatter):
         self.result = fitresult
         self.dataset_used = dataset_used
 
-
-    def printscatter(self, fit, mima=None, pointsize=0.5,
-                     regression=True, newcalc=False,  **kwargs):
-        '''
-        geerate a scatterplot to investigate the quality of the fit
-
-        Parameters:
-        ------------
-        fit : list
-              output of monofit()-function
-        Other Parameters:
-        ------------------
-        mima : list
-               manual definition plot-boundaries via mima = [min, max]
-        pointsize : float
-                    manual specification of pointsize
-        regression : bool (default = True)
-                     indicator if the scipy.stats.linregress should be called
-                     to get the regression-line and the r^2 value
-        kwargs : -
-                 kwargs passed to matplotlib.pyplot.scatter()
-        '''
-
-        (res_lsq, R, data, inc, mask, weights,
-         res_dict, start_dict, fixed_dict) = fit
-
-        # reset incidence-angles in case they have been altered beforehand
-        R.t_0 = inc
-        R.p_0 = np.zeros_like(inc)
-
-        fig = plt.figure()
-        ax = fig.add_subplot(111)
-
-        if newcalc is True:
-
-            # for python > 3.4
-            # calc_dict = dict(**res_dict, **fixed_dict)
-            calc_dict = dict((k, v) for k, v in list(res_dict.items())
-                             + list(fixed_dict.items()))
-
-            estimates = self._calc_model(R, calc_dict)
-
-            # apply mask
-            estimates = estimates[~mask]
-            measures = data[~mask]
-
-        else:
-            # get the residuals and apply mask
-            residuals = np.reshape(res_lsq.fun, data.shape)
-            residuals = np.ma.masked_array(residuals, mask)
-            # prepare measurements
-            measures = data[~mask]
-            # calculate estimates
-            estimates = residuals[~mask] + measures
-
-        if mima is None:
-            mi = np.min((measures, estimates))
-            ma = np.max((measures, estimates))
-        else:
-            mi, ma = mima
-
-        ax.scatter(estimates, measures, s=pointsize, alpha=0.7, **kwargs)
-
-        # plot 45degree-line
-        ax.plot([mi, ma], [mi, ma], 'k--')
-
-        if self.sig0 is True:
-            quantity = r'$\sigma_0$'
-        else:
-            quantity = 'Intensity'
-
-        if self.dB is True:
-            scale = '[dB]'
-        else:
-            scale = ''
-
-        ax.set_xlabel('modelled ' + quantity + scale)
-        ax.set_ylabel('measured ' + quantity + scale)
-
-        if regression is True:
-            # evaluate linear regression to get r-value etc.
-            slope, intercept, r_value, p_value, std_err = linregress(estimates,
-                                                                     measures)
-
-            ax.plot(np.sort(measures),
-                    intercept + slope * np.sort(measures), 'r--', alpha=0.4)
-
-            ax.text(0.8, .1, '$R^2$ = ' + str(np.round(r_value**2, 2)),
-                    horizontalalignment='center',
-                    verticalalignment='center',
-                    transform=ax.transAxes)
-
-        return fig, r_value**2
-
-
-    def printsingle(self, fit, measurements=None, datelist=None,
-                    dates=None, hexbinQ=True, hexbinargs={},
-                    convertTodB=False):
-        '''
-        a function to investigate the quality of the individual fits
-
-        Parameters:
-        ------------
-        fit : list
-              output of the monofit()-function
-        measurements : list
-                       a list containing the number of the measurement
-                       that should be plotted (starting from 1)
-
-        Other Parameters:
-        ------------------
-        datelist : list
-                   a list used to label the x-axis
-                   the shape must be [ groups, indicators ]
-                   where groups is a list of G values that are used to
-                   group the measurements, and
-                   indicators is a list containing N elements grouped into
-                   G groups, where N is the number of measurements.
-
-                   e.g. if you had a measurement in January and February
-                   for the years 2014 and 2015, you have:
-                       datelist = [[2014, 2015], [ [1,2],[1,2]] ]
-        dates : list
-                if datelist has been provided, the dates-list can be used to
-                select the measurement that shall be plotted.
-                The shape is : [[group_0, indicator_0],
-                                [group_1, indicator_1],
-                                ...]
-
-                using the example for the datelist above, the second and fourth
-                measurement can be selected via:
-                    dates = [[2014, 2], [2015, 2]]
-
-        hexbinQ : bool (default = False)
-                  indicator if a hexbin-plot should be underlayed
-                  that shows the distribution of the datapoints
-
-        hexbinargs : dict
-                     a dict containing arguments to customize the hexbin-plot
-        convertTodB : bool (default=False)
-                      if set to true, the datasets will be converted to dB
-        '''
-
-        from matplotlib.colors import LinearSegmentedColormap
-        from matplotlib.colors import Normalize
-
-        # function to generate colormap that fades between colors
-        def CustomCmap(from_rgb, to_rgb):
-
-            # from color r,g,b
-            r1, g1, b1 = from_rgb
-
-            # to color r,g,b
-            r2, g2, b2 = to_rgb
-
-            cdict = {'red': ((0, r1, r1),
-                             (1, r2, r2)),
-                     'green': ((0, g1, g1),
-                               (1, g2, g2)),
-                     'blue': ((0, b1, b1),
-                              (1, b2, b2))}
-
-            cmap = LinearSegmentedColormap('custom_cmap', cdict)
-            return cmap
-
-        (res_lsq, R, data, inc, mask, weights,
-         res_dict, start_dict, fixed_dict) = fit
-
-        # reset incidence-angles in case they have been altered beforehand
-        R.t_0 = inc
-        R.p_0 = np.zeros_like(inc)
-
-        # for python > 3.4
-        # calc_dict = dict(**res_dict, **fixed_dict)
-        calc_dict = dict((k, v) for k, v in list(res_dict.items())
-                         + list(fixed_dict.items()))
-
-        estimates = self._calc_model(R, calc_dict)
-
-        if datelist is None and dates is not None:
-            assert False, 'you can only provide dates if dateslist is provided'
-
-        if dates is not None:
-            measurements = []
-
-            for dd in dates:
-                yval = np.argwhere(np.array(datelist[0]) == dd[0])[0][0]
-                try:
-                    mval = np.argwhere(np.array(datelist[1][yval]) ==
-                                       dd[1])[0][0]
-                    mval = (mval +
-                            int(np.sum([len(i)for i in datelist[1][:yval]])))
-
-                    measurements = measurements + [mval]
-                except Exception:
-                    # TODO this is a stupid way to handle this
-                    assert False, 'There is something wrong with the dates'
-        else:
-            # since python starts counting at 0 ...
-            measurements = np.array(measurements) - 1
-
-        fig = plt.figure()
-        ax = fig.add_subplot(111)
-
-        for m_i, m in enumerate(measurements):
-
-            if convertTodB is True:
-                y = 10.*np.log10(estimates[m][~mask[m]])
-            else:
-                y = estimates[m][~mask[m]]
-
-            # plot data
-            if datelist is None:
-                label = m + 1
-            else:
-                monlen = 0
-                for i, mon in enumerate(datelist[1]):
-                    monlen = monlen + len(mon)
-                    if m < monlen:
-                        labelyear = datelist[0][i]
-                        break
-
-                label = (str(labelyear) + ' ' +
-                         str(np.concatenate(datelist[1])[m]))
-
-            xdata = np.rad2deg(inc[m][~mask[m]])
-
-            if convertTodB is True:
-                ydata = 10.*np.log10(data[m][~mask[m]])
-            else:
-                ydata = data[m][~mask[m]]
-
-            # get color that will be applied to the next line drawn
-            dummy, = ax.plot(xdata[0], ydata[0], '.', alpha=0.)
-            color = dummy.get_color()
-
-            if hexbinQ is True:
-                args = dict(gridsize=15, mincnt=1,
-                            linewidths=0., vmin=0.5, alpha=0.7)
-                args.update(hexbinargs)
-
-                # evaluate the hexbinplot once to get the maximum number of
-                # datapoints within a single hexagonal (used for normalization)
-                dummyargs = args.copy()
-                dummyargs.update({'alpha': 0.})
-                hb = ax.hexbin(xdata, ydata, **dummyargs)
-
-                # generate colormap that fades from white to the color
-                # of the plotted data  (asdf.get_color())
-                cmap = CustomCmap([1.00, 1.00, 1.00],
-                                  plt.cm.colors.hex2color(color))
-                # setup correct normalizing instance
-                norm = Normalize(vmin=0, vmax=hb.get_array().max())
-
-                ax.hexbin(xdata, ydata, cmap=cmap, norm=norm, **args)
-
-            # plot datapoints
-            asdf, = ax.plot(xdata, ydata, '.',
-                            color=color, alpha=1.,
-                            label=label, markersize=10)
-
-            # plot results
-            iii = inc[m][~mask[m]]
-            ax.plot(np.rad2deg(iii[np.argsort(iii)]), y[np.argsort(iii)],
-                    '-', color='w', linewidth=3)
-
-            ax.plot(np.rad2deg(iii[np.argsort(iii)]), y[np.argsort(iii)],
-                    '-', color=asdf.get_color(), linewidth=2)
-
-            ax.set_xlabel('$\\theta_0$ [deg]')
-            ax.set_ylabel('$\\sigma_0$ [dB]')
-
-        if dates is None:
-            ax.legend(title='# Measurement')
-            # plt.setp(legend.get_title(),fontsize=8) # change fontsize
-        else:
-            ax.legend(title='# Date')
-
-        return fig
-
-    def printseries(self, fit, index=None, legends=True, minmax=None,
-                    newcalc=False, convertTodB=False):
-        '''
-        a function to quickly print the fit-results and the gained parameters
-
-        Parametsrs:
-        ------------
-        fit : list
-              output of monofit_all()-function
-        datelist : list
-                   a list used to label the x-axis
-                   the shape must be [ groups, indicators ]
-                   where groups is a list of G values that are used to
-                   group the measurements, and
-                   indicators is a list containing N elements grouped into
-                   G groups, where N is the number of measurements.
-
-                   e.g. if you had a measurement in January and February
-                   for the years 2014 and 2015, you have:
-                       datelist = [[2014, 2015], [[1,2],[1,2]] ]
-        legends : bool (default = True)
-                  indicator if legends should be plotted
-
-        Returns:
-        ---------
-        fig : matplotlib.figure object
-        '''
-
-        (res_lsq, R, data, inc, mask, weights,
-         res_dict, start_dict, fixed_dict) = fit
-
-        if minmax is None:
-            minmax = [0, len(data)]
-
-        if newcalc is True:
-
-            # for python > 3.4
-            # calc_dict = dict(**res_dict, **fixed_dict)
-            calc_dict = dict((k, v) for k, v in list(res_dict.items())
-                             + list(fixed_dict.items()))
-
-            estimates = self._calc_model(R, calc_dict)
-        else:
-            # get the residuals and apply mask
-            residuals = np.reshape(res_lsq.fun, data.shape)
-            residuals = np.ma.masked_array(residuals, mask)
-            # prepare measurements
-            measures = np.ma.masked_array(data, mask)
-            # calculate estimates
-            estimates = residuals + measures
-
-        maskedestimates = np.ma.masked_array(estimates,
-                                             mask)[minmax[0]:minmax[1]]
-
-        # get number of measurement as index
-        Nvals = np.ones_like(maskedestimates,
-                             dtype=int) * (np.arange(*minmax))[:, np.newaxis]
-        Nvals = np.ma.concatenate(Nvals)
-        sorts = np.ma.argsort(Nvals)
-
-        maskedestimates = np.concatenate(maskedestimates)
-        maskeddata = np.ma.concatenate(
-                np.ma.masked_array(data, mask)[minmax[0]:minmax[1]])
-
-        if convertTodB is True:
-            maskedestimates = 10.*np.ma.log10(maskedestimates)
-            maskeddata = 10.*np.ma.log10(maskeddata)
-
-        fig = plt.figure()
-        ax = fig.add_subplot(111)
-
-        if index is None:
-            Nmeasurements = len(inc)
-            ax.set_xticks(np.arange(1, Nmeasurements + 1))
-            ax.set_xlabel('# Measurement')
-
-            ax.plot(Nvals[sorts], maskeddata[sorts], '-',
-                    label='data', marker='.')
-            ax.plot(Nvals[sorts], maskedestimates[sorts], '-',
-                    label='model', marker='.', alpha=0.5)
-
-        else:
-            import matplotlib.dates as dates
-            from datetime import timedelta
-
-            fullindex = pd.DatetimeIndex(
-                np.concatenate([[i] * len(data[0]) for i in index.values]))
-
-            ax.plot(fullindex,
-                    np.ma.concatenate(np.ma.masked_array(data, mask)),
-                    '-', label='data', marker='.')
-
-            ax.plot(fullindex,
-                    np.ma.concatenate(np.ma.masked_array(estimates, mask)),
-                    '-', label='model', marker='.', alpha=0.5)
-
-            ax.set_xlim([index[0] - timedelta(days=20),
-                         index[-1] + timedelta(days=20)])
-
-            ax.xaxis.set_minor_locator(dates.MonthLocator())
-            ax.xaxis.set_minor_formatter(dates.DateFormatter('%m'))
-
-            ax.xaxis.set_major_locator(dates.YearLocator())
-            ax.xaxis.set_major_formatter(dates.DateFormatter('\n%Y'))
-
-        ax.set_xlabel('Measurements')
-        ax.set_ylabel('$\\sigma_0$ [dB]')
-
-        if legends is True:
-            plt.legend()
-
-        return fig
-
-    def printviolin(self, inputs, param, together=True,
-                    func=np.mean, print_mean=False,
-                    bw_method=None, funcargs={}, **kwargs):
-        '''
-        plot a violin-plot of defined fit-result parameters
-
-        Parameters:
-        ------------
-        inputs : dict
-                 a dictionary containing the plotlabels(keys) and a list of
-                 results of the monofit-function (values) that are intended
-                 to be compared, e.g.:
-                     inputs = {'name1' : [monofit1, monofit2, ...],
-                               'name2' : [monofit3, monofit4, ...]'}
-        param : str
-                the name of the parameter to be plotted (the corresponding key
-                of the results dictionary returned by the monofit function)
-        together : bool
-                   indicator whether or not a combined violin shall be plotted
-        func : function-call signature (default = np.mean)
-               the call-signature of a function that will be applied to the
-               list of fitted-values for the chosen parameter
-        print_mean : bool
-                     indicator whether or not dotted lines for the mean-values
-                     shall be plotted
-        bw_method : str, int or None
-                    selector for the chosen method for estimating the bandwidth
-                    in the generation of the violinplot
-                    (parameter of matplotlib.pyplot.violin)
-        **kwargs : kwargs
-                   keyword-arguments forwraded to matplotlib.pyplot.figure()
-
-        Returns:
-        ---------
-        fig : matplotlib.figure
-        '''
-
-        fig = plt.figure(**kwargs)
-        ax = fig.add_subplot(111)
-        resval = {}
-
-        colors = ['r', 'g', 'b', 'c', 'm', 'y']
-
-        for reskey in inputs:
-            inp = inputs[reskey]
-            # evaluate sortpattern to sort results based on coverage
-            # p, c = sfc.printfracs(inp)
-            # plt.close(p)
-            # sortpa[reskey] = np.argsort(np.sum(c, axis=0))
-
-            val = []
-
-            if param == 'residual':
-                # estimate and print rel. residuals of the test_i measurement
-                rel_residual = []
-                for test_i in np.array(inp):
-                    # get results of i'th measurement of fit-result inp
-                    (res_lsq, _, data, _, mask, weights,
-                     _, _, _) = test_i[0]
-
-                    # get the residuals and apply mask
-                    residuals = np.reshape(res_lsq.fun, data.shape)
-                    residuals = np.ma.masked_array(residuals, mask)
-                    # prepare measurements
-                    measures = np.ma.masked_array(data, mask)
-                    # calculate estimates
-                    estimates = residuals + measures
-                    # calculate relative absolute residuals
-                    rel_residual = np.ma.abs(residuals/estimates)
-                    # calculate mean relative residual over incidence-angles
-                    mean_rel_residual = np.ma.mean(rel_residual, axis=1)
-                    # convert result to ordinary list
-                    mean_rel_residual = list(mean_rel_residual)
-
-                    # apply function
-                    addval = func(mean_rel_residual, **funcargs)
-                    if np.isscalar(addval):
-                        addval = [addval]
-                    else:
-                        addval = list(addval)
-                    val += addval
-            else:
-                # extract res_dict from given result
-                params = [i[6] for i in np.array(inp)[:, 0]]
-
-                for i in params:
-                    addval = func(i[param], **funcargs)
-                    if np.isscalar(addval):
-                        addval = [addval]
-                    else:
-                        addval = list(addval)
-                    val += addval
-
-            resval[reskey] = np.array(val)
-
-        # print each result as an individual violin plot
-        vio_ind = ax.violinplot(resval.values(),
-                                showmeans=True,
-                                showextrema=True,
-                                bw_method=bw_method)
-        # now change colors
-        i = 0
-        for pc in vio_ind['bodies']:
-            pc.set_facecolor(colors[i % len(colors)])
-            pc.set_edgecolor(colors[i % len(colors)])
-            pc.set_alpha(.8)
-            i = i + 1
-
-        for i in ['cmeans', 'cmins', 'cmaxes', 'cbars']:
-            vio_ind[i].set_color('k')
-            vio_ind[i].set_alpha(0.5)
-
-        # set labels
-        labels = [i for i in resval.keys()]
-        nticks = len(vio_ind['bodies']) + 1
-
-        if together is True:
-            # print the violinplots together
-            # (with only the outer line being plotted)
-            i = 0
-            for key in resval:
-                vio = ax.violinplot(resval[key],
-                                    positions=[len(vio_ind['bodies']) + 1],
-                                    showmeans=False,
-                                    showextrema=False,
-                                    bw_method=bw_method)
-                # now change colors
-                for pc in vio['bodies']:
-                    pc.set_facecolor('None')
-                    pc.set_edgecolor(colors[i % len(colors)])
-                    pc.set_alpha(1.)
-                    pc.set_linewidth(1.)
-                i = i + 1
-
-            # print a shading for the violinplots
-            i = 0
-            for key in resval:
-                vio = ax.violinplot(resval[key],
-                                    positions=[len(vio_ind['bodies']) + 1],
-                                    showmeans=False,
-                                    showextrema=False,
-                                    bw_method=bw_method)
-                # now change colors
-                for pc in vio['bodies']:
-                    pc.set_facecolor(colors[i % len(colors)])
-                    pc.set_edgecolor('None')
-                    pc.set_alpha(.2)
-                i = i + 1
-
-            labels += ['all']
-            nticks += 1
-
-        ax.set_xticks(np.arange(1, nticks))
-        ax.set_xticklabels(labels)
-
-        if print_mean is True:
-            i = 0
-            for key in resval:
-                ax.plot([0.5, nticks - 0.5],
-                        [np.mean(resval[key])] * 2,
-                        linestyle='--',
-                        color=colors[i % len(colors)],
-                        alpha=0.5,
-                        linewidth=0.5)
-                i += 1
-
-        if param == 'residual':
-            ax.set_title('Parameter: relative residual')
-            ax.set_ylabel('relative residual')
-        else:
-            ax.set_title('Parameter: ' + str(param))
-            ax.set_ylabel(str(param))
-
-        return fig
