+++ rt1/rtplots.py
@@ -17,6 +17,9 @@ except ImportError:
 import matplotlib as mpl
 import matplotlib.pyplot as plt
 import matplotlib.lines as mlines
+from matplotlib.colors import LinearSegmentedColormap
+from matplotlib.colors import Normalize
+
 # plot of 3d scattering distribution
 #import mpl_toolkits.mplot3d as plt3d
 
@@ -367,237 +370,6 @@ def polarplot(R=None, SRF=None, V=None, incp=[15., 35., 55., 75.],
     plt.show()
     return polarfig
 
-def logmono(inc, Itot=None, Isurf=None, Ivol=None, Iint=None,
-            ylim=None, sig0=False, noint=False, fractions=True,
-            label=None, label_frac=None):
-    """
-    Generate a plot of the monostatic backscattered Intensity or sigma_0
-    in dB as well as a plot showing the fractional contributions to the
-    total signal.
-
-    Parameters
-    -----------
-    inc : float-array
-          Incidence-angle range used for calculating the intensities
-    Itot, Ivol, Isurf, Iint : array_like(float)
-                              individual monostatic signal contributions
-                              i.e. outputs from RT1.calc()  with
-                              RT1.geometry = 'mono'  and    RT1.t_0 = inc
-                              At least one of the arrays must be provided
-                              and it must be of the same length as inc!
-    Other Parameters
-    -----------------
-    ylim : [float , float]
-           Manual entry of plot-boundaries as [ymin, ymax]
-    sig0 : boolean (default = False)
-           Indicator to choose whether sigma_0 (True) or intensity (False)
-           is plotted. If True, the relation
-           sigma_0 = 4 * Pi * cos(inc) * intensity    is applied.
-    noint : boolean (default = False)
-            If True, the zero-order contribution
-            (i.e. I_tot0 = I_vol + I_surf) is plotted as a dashed line
-    fractions : boolean (default = True)
-                If True, a plot of the fractional contributions to the
-                total signal is generated.
-                (i.e. volume_fraction = I_vol / I_tot    etc.)
-    label : string
-            Manual label of backscatter-plot
-    label_frac : string
-            Manual label of the fractional contribution-plot
-
-    Returns
-    -------
-    f : figure
-        a matplotlib figure showing a log-plot of the given input-arrays
-    """
-
-    # define a mask to avoid log(0) errors (used to crop numpy arrays)
-    def mask(x):
-        return (x <= 0.)
-
-    assert isinstance(inc, np.ndarray), 'Error, inc must be a numpy array'
-
-    if Itot is not None:
-        assert isinstance(Itot, np.ndarray), 'Error, Itot must be a ' + \
-            'numpy array'
-        assert len(inc) == len(Itot), 'Error: Length of inc and Itot ' + \
-            'is not equal'
-        Itot = np.ma.array(Itot, mask=mask(Itot))
-
-    if Isurf is not None:
-        assert isinstance(Isurf, np.ndarray), 'Error, Isurf must be a ' + \
-            'numpy array'
-        assert len(inc) == len(Isurf), 'Error: Length of inc and ' + \
-            'Isurf is not equal'
-        Isurf = np.ma.array(Isurf, mask=mask(Isurf))
-
-    if Ivol is not None:
-        assert isinstance(Ivol, np.ndarray), 'Error, Ivol must be a ' + \
-            'numpy array'
-        assert len(inc) == len(Ivol), 'Error: Length of inc and Ivol' + \
-            ' is not equal'
-        Ivol = np.ma.array(Ivol, mask=mask(Ivol))
-
-    if Iint is not None:
-        assert isinstance(Iint, np.ndarray), 'Error, Iint must be' + \
-            ' a numpy array'
-        assert len(inc) == len(Iint), 'Error: Length of inc and Iint' + \
-            ' is not equal'
-        Iint = np.ma.array(Iint, mask=mask(Iint))
-
-    if label is not None:
-        assert isinstance(label, str), 'Error, Label must be a string'
-
-    if ylim is not None:
-        assert len(ylim) == 2, 'Error: ylim must be an array ' + \
-            'of length 2!   ylim = [ymin, ymax]'
-    if ylim is not None:
-        assert isinstance(ylim[0], (int, float)), 'Error: ymin must' + \
-            ' be a number'
-    if ylim is not None:
-        assert isinstance(ylim[1], (int, float)), 'Error: ymax must' + \
-            ' be a number'
-
-    if noint is True:
-        assert Isurf is not None, 'Isurf must be provided if noint = True'
-        assert Ivol is not None, 'Ivol must be provided if noint = True'
-
-    assert isinstance(sig0, bool), 'Error: sig0 must be True or False'
-    assert isinstance(fractions, bool), 'Error: fractions must' + \
-        ' be either True or False'
-
-    ctot = 'black'
-    csurf = 'red'
-    cvol = 'green'
-    cint = 'blue'
-
-    if sig0 is True:
-        #  I..  will be multiplied with signorm to get sigma0 values
-        # instead of normalized intensity
-        signorm = 4. * np.pi * np.cos(np.deg2rad(inc))
-    else:
-        signorm = 1.
-
-    if fractions is True:
-        f = plt.figure(figsize=(14, 7))
-        ax = f.add_subplot(121)
-        ax2 = f.add_subplot(122)
-    else:
-        f = plt.figure(figsize=(7, 7))
-        ax = f.add_subplot(111)
-
-    ax.grid()
-    ax.set_xlabel('$\\theta_0$ [deg]')
-
-    if sig0 is True:
-
-        if Itot is not None:
-            ax.plot(inc, 10. * np.log10(signorm * Itot),
-                    color=ctot, label='$\\sigma_0^{tot}$')
-        if Isurf is not None:
-            ax.plot(inc, 10. * np.log10(signorm * Isurf),
-                    color=csurf, label='$\\sigma_0^{surf}$')
-        if Ivol is not None:
-            ax.plot(inc, 10. * np.log10(signorm * Ivol),
-                    color=cvol, label='$\\sigma_0^{vol}$')
-        if Iint is not None:
-            ax.plot(inc, 10. * np.log10(signorm * Iint),
-                    color=cint, label='$\\sigma_0^{int}$')
-        if noint is True:
-            ax.plot(inc,
-                    10. * np.log10(signorm * (Ivol + Isurf)),
-                    color=ctot, linestyle='--',
-                    label='$\\sigma_0^{surf}+\\sigma_0^{vol}$')
-
-        if label is None:
-            ax.set_title('Bacscattering Coefficient')
-        else:
-            ax.set_title(label)
-
-        ax.set_ylabel('$\\sigma_0$ [dB]')
-
-    else:
-
-        if Itot is not None:
-            ax.plot(inc, 10. * np.log10(signorm * Itot),
-                    color=ctot, label='$I_{tot}$')
-        if Isurf is not None:
-            ax.plot(inc, 10. * np.log10(signorm * Isurf),
-                    color=csurf, label='$I_{surf}$')
-        if Ivol is not None:
-            ax.plot(inc, 10. * np.log10(signorm * Ivol),
-                    color=cvol, label='$I_{vol}$')
-        if Iint is not None:
-            ax.plot(inc, 10. * np.log10(signorm * Iint),
-                    color=cint, label='$I_{int}$')
-        if noint is True:
-            ax.plot(inc, 10. * np.log10(signorm * (Ivol + Isurf)),
-                    color=ctot, linestyle='--',
-                    label='$I_0^{surf}+I_0^{vol}$')
-
-        if label is None:
-            ax.set_title('Normalized Intensity')
-        else:
-            ax.set_title(label)
-
-        ax.set_ylabel('$I^+$ [dB]')
-    legend = ax.legend()
-    legend.get_frame().set_facecolor('w')
-    legend.get_frame().set_alpha(.5)
-
-    if ylim is None:
-        Itotmax = np.nan
-        Isurfmax = np.nan
-        Ivolmax = np.nan
-        Iintmax = np.nan
-
-        # set minimum y to the smallest value of the maximas of
-        # the individual contributions -5.
-        if Itot is not None:
-            Itotmax = np.nanmax(10. * np.log10(signorm * Itot))
-        if Isurf is not None:
-            Isurfmax = np.nanmax(10. * np.log10(signorm * Isurf))
-        if Ivol is not None:
-            Ivolmax = np.nanmax(10. * np.log10(signorm * Ivol))
-        if Iint is not None:
-            Iintmax = np.nanmax(10. * np.log10(signorm * Iint))
-
-        ymin = np.nanmin([Itotmax, Isurfmax, Ivolmax, Iintmax]) - 5.
-
-        # set maximum y to the maximum value of the maximas
-        # of the individual contributions + 5.
-        ymax = np.nanmax([Itotmax, Isurfmax, Ivolmax, Iintmax]) + 5.
-
-        ax.set_ylim(ymin, ymax)
-    else:
-        ax.set_ylim(ylim[0], ylim[1])
-
-    if fractions is True:
-        # plot fractions
-        if Itot is not None and Isurf is not None:
-            ax2.plot(inc, Isurf / Itot, label='surface', color=csurf)
-        if Itot is not None and Ivol is not None:
-            ax2.plot(inc, Ivol / Itot, label='volume', color=cvol)
-        if Itot is not None and Iint is not None:
-            ax2.plot(inc, Iint / Itot, label='interaction', color=cint)
-
-        if label_frac is None:
-            ax2.set_title('Fractional contributions to total signal')
-        else:
-            ax2.set_title(label_frac)
-
-        ax2.set_xlabel('$\\theta_0$ [deg]')
-        if sig0 is True:
-            ax2.set_ylabel('$\\sigma_0 / \\sigma_0^{tot}$')
-        else:
-            ax2.set_ylabel('$I / I_{tot}$')
-        ax2.grid()
-        legend2 = ax2.legend()
-        legend2.get_frame().set_facecolor('w')
-        legend2.get_frame().set_alpha(.5)
-
-    plt.show()
-    return f
 
 def linplot3d(theta, phi, Itot=None, Isurf=None, Ivol=None,
               Iint=None, surfmultip=1., zoom=2.):
@@ -896,553 +668,872 @@ def hemreflect(R=None, SRF=None, phi_0=0., t_0_step=5., t_0_min=0.,
         plt.show()
         return fig
 
-def printsig0timeseries(fit,
-                        dB=True,
-                        sig0=True,
-                        params=None,
-                        printtot = True,
-                        printsurf = True,
-                        printvol = True,
-                        printint = True,
-                        printorig = True,
-                        months = None,
-                        years = None,
-                        ylim=None,
-                        printinc = True):
+
+class plots:
     '''
-    Print individual contributions, resulting parameters and the
-    reference dataset of an rt1.rtfits object as timeseries.
+    Generation of plots to visualize rtfits results
 
-    Parameters:
-    -------------
-    fit : rtfits object
-          the rtfits-object containing the fit-results
-    dB : bool (default = True)
-         indicator if the plot is intended to be in dB or linear units
-    sig0 : bool (default = True)
-         indicator if the plot should display sigma_0 (sig0) or intensity (I)
-         The applied relation is: sig0 = 4.*pi*cos(theta) * I
-    params: list
-            a list of parameter-names that should be overprinted
-            (the names must coincide with the arguments of set_V_SRF())
-    printtot, printsurf, printvol, printint, printorig : bool
-            indicators if the corresponding components should be plotted
-    months : list (default = None)
-             a list of months to plot (if None, all will be plotted)
-    years : list (default = None)
-            a list of years to select (if None, all will be plotted)
-    ylim : tuple
-           a tuple of (ymin, ymax) that will be used as boundaries for the
-           y-axis
-    printinc : bool (default = True)
-               indicator if the incidence-angle dependency should be plotted
-               (in a separate plot alongside the timeseries)
-    '''
-    # get mask
-    (_, _, data, _, mask, _, _, _, _) = fit.result
-    # get incidence-angles
-    inc_array = np.ma.masked_array(fit.result[1].t_0, mask)
-    inc = inc_array.compressed()
-    # get input dataset
-    data = np.ma.masked_array(data, mask)
+    - printscatter
+        scatterplot of measured vs. modelled data
 
-    def dBsig0convert(val):
-        # if results are provided in dB convert them to linear units
-        if fit.dB is True: val = 10**(val/10.)
-        # convert sig0 to intensity
-        if sig0 is False and fit.sig0 is True:
-            val = val/(4.*np.pi*np.cos(inc))
-        # convert intensity to sig0
-        if sig0 is True and fit.sig0 is False:
-            val = 4.*np.pi*np.cos(inc)*val
-        # if dB output is required, convert to dB
-        if dB is True: val = 10.*np.log10(val)
-        return val
+    - printresults
 
-    res_dict = {}
-    # add fitted parameters
-    res_dict.update(fit.result[6])
-    # add constant values
-    res_dict.update(fit.result[-1])
-
-    # calculate individual contributions
-    contrib_array = fit._calc_model(R=fit.result[1],
-                                    res_dict=res_dict,
-                                    return_components=True)
-
-    # apply mask and convert to pandas dataframe
-    contrib_array = [np.ma.masked_array(con, mask) for con in contrib_array]
-    contrib_array += [data, inc_array]
-
-    contrib = []
-    for i, cont in enumerate(contrib_array):
-        contrib += [pd.concat([pd.DataFrame(i, index = fit.index) for i in cont.T])[0]]
-
-    contrib = pd.concat(contrib,
-                        keys=['tot', 'surf', 'vol', 'inter',
-                              '$\\sigma_0$ dataset', 'inc'], axis=1).dropna()
-
-    # convert units
-    contrib[['tot', 'surf', 'vol', 'inter',
-             '$\\sigma_0$ dataset']] = contrib[[
-                     'tot', 'surf', 'vol', 'inter', '$\\sigma_0$ dataset'
-                     ]].apply(dBsig0convert)
-
-    # drop unneeded columns
-    if fit.result[1].int_Q is False or printint is False:
-        contrib = contrib.drop('inter', axis=1)
-    if printtot is False: contrib = contrib.drop('tot', axis=1)
-    if printsurf is False: contrib = contrib.drop('surf', axis=1)
-    if printvol is False: contrib = contrib.drop('vol', axis=1)
-    if printorig is False:
-        contrib = contrib.drop('$\\sigma_0$ dataset', axis=1)
-
-    # select years and months
-    if years is not None:
-        contrib = contrib.loc[contrib.index.year.isin(years)]
-    if months is not None:
-        contrib = contrib.loc[contrib.index.month.isin(months)]
-
-    # print incidence-angle dependency
-    if printinc is True:
-        f, [ax, ax_inc] = plt.subplots(ncols=2, figsize=(15,5),
-                                       gridspec_kw={'width_ratios':[3,1]},
-                                       sharey=True)
-        f.subplots_adjust(left=0.05, right=0.98, top=0.98,
-                          bottom=0.1, wspace=0.1)
-
-        # -------------------
-        color = {'tot':'r', 'surf':'b', 'vol':'g', 'inter':'y',
-                 '$\\sigma_0$ dataset':'k'}
-
-        groupedcontrib = contrib.groupby(contrib.index)
-
-        #return contrib, groupedcontrib
-        for label in contrib.keys():
-            if label in ['inc']: continue
-            a=np.rad2deg(rectangularize([x.values for _, x in groupedcontrib['inc']])).T
-            b=np.array(rectangularize([x.values for _, x in groupedcontrib[label]])).T
-            x = (np.array([a,b]).T)
-
-            l_col = mpl.collections.LineCollection(x,linewidth =.25, label='x',
-                                      color=color[label], alpha = 0.5)
-            ax_inc.add_collection(l_col)
-            ax_inc.scatter(a, b, color=color[label], s=1)
-            ax_inc.set_xlim(a.min(), a.max())
-            ax_inc.set_xlabel('$\\theta_0$')
-        ax_inc.set_xlabel('$\\theta_0$')
-
-
-    else:
-        f, ax = plt.subplots(figsize=(12,5))
-        f.subplots_adjust(left=0.05, right=0.98, top=0.98,
-                          bottom=0.1, wspace=0.05)
-
-    for label, val in contrib.items():
-        if label in ['inc']: continue
-        color = {'tot':'r', 'surf':'b', 'vol':'g', 'inter':'y'}
-        if printorig is True: color['$\\sigma_0$ dataset'] = 'k'
-        ax.plot(val.sort_index(), linewidth =.25, marker='.',
-                ms=2, label=label, color=color[label], alpha = 0.5)
-    # overprint parameters
-    if params != None:
-        paramdf_dict = {}
-        # add fitted parameters
-        paramdf_dict.update(fit.result[6])
-        # add constant values
-        paramdf_dict.update(fit.result[-1])
-
-        paramdf = pd.DataFrame(paramdf_dict,
-                               index = fit.index).sort_index()
-        if years is not None:
-            paramdf = paramdf.loc[paramdf.index.year.isin(years)]
-        if months is not None:
-            paramdf = paramdf.loc[paramdf.index.month.isin(months)]
-
-        pax = ax.twinx()
-        for k in params:
-            pax.plot(paramdf[k], lw=1, marker='.', ms=2)
-        pax.legend(loc='upper right', ncol=5)
-        pax.set_ylabel('parameter-values')
-
-    # format datetime index
-    ax.xaxis.set_minor_locator(mpl.dates.MonthLocator())
-    ax.xaxis.set_minor_formatter(mpl.dates.DateFormatter('%m'))
-    ax.xaxis.set_major_locator(mpl.dates.YearLocator())
-    ax.xaxis.set_major_formatter(mpl.dates.DateFormatter('\n%Y'))
-
-    # set ylabels
-    if sig0 is True:
-        label = '$\\sigma_0$'
-    else:
-        label = 'Intensity'
-    if dB is True: label += ' [dB]'
-    ax.set_ylabel(label)
-
-    # generate legend
-    hand, lab = ax.get_legend_handles_labels()
-    lab, unique_ind = np.unique(lab, return_index=True)
-    ax.legend(handles = list(np.array(hand)[unique_ind]),
-              labels=list(lab), loc='upper left',
-              ncol=5)
-
-    if ylim is not None:
-        ax.set_ylim(ylim)
-
-
-
-def plot_interres(fit, dynparam = ['SM'],
-                  cmaps=['viridis_r', 'RdBu','inferno', 'magma_r']):
-    '''
-    a function to plot the intermediate-results
-    (the data is only available if rtfits.monofit has been called with
-    the argument intermediate_results=True)
-
-    Parameters:
-    -------------
-    fit : rtfits object
-          the rtfits-object containing the fit-results
-
-    dynparam : a list of parameter-names that are intended to be plotted
-               as timeseries.
-    '''
-    interparams = {}
-    for i, valdict in enumerate(fit.intermediate_results['parameters']):
-        for key, val in valdict.items():
-            if key in interparams:
-                interparams[key] += [[i, np.mean(val)]]
-            else:
-                interparams[key] = [[i, np.mean(val)]]
-    intererrs = {}
-    for i, valdict in enumerate(fit.intermediate_results['residuals']):
-        for key, val in valdict.items():
-            if key in intererrs:
-                intererrs[key] += [[i, np.mean(val)]]
-            else:
-                intererrs[key] = [[i, np.mean(val)]]
-    interjacs = {}
-    for i, valdict in enumerate(fit.intermediate_results['jacobian']):
-        for key, val in valdict.items():
-            if key in interjacs:
-                interjacs[key] += [[i, np.mean(val)]]
-            else:
-                interjacs[key] = [[i, np.mean(val)]]
-
-    for i in [interparams, intererrs, interjacs]:
-        for key, val in i.items():
-            i[key] = np.array(val).T
-
-    interres_params = {}
-    for key in dynparam:
-        interres_p = pd.concat([pd.DataFrame(valdict[key],
-                                             fit.index.drop_duplicates(),
-                                columns=[i])
-                                for i, valdict in enumerate(
-                                        fit.intermediate_results['parameters'])
-                                ], axis=1)
-        interres_params[key] = interres_p
-
-
-    f = plt.figure(figsize=(15,10))
-    f.subplots_adjust(top=0.98, left=0.05, right=0.95)
-    gs = mpl.gridspec.GridSpec(4, len(interjacs),
-                           #width_ratios=[1, 2],
-                           #height_ratios=[1, 2, 1]
-                           )
-    axsm = plt.subplot(gs[0,:])
-    axerr = plt.subplot(gs[1,:])
-    paramaxes, jacaxes = [], []
-    for i in range(len(interjacs)):
-        paramaxes += [plt.subplot(gs[2,i])]
-        jacaxes += [plt.subplot(gs[3,i])]
-
-    smhandles, smlabels = [],[]
-    for nparam, [parameter, paramdf] in enumerate(interres_params.items()):
-
-        cmap = plt.get_cmap(cmaps[nparam])
-
-        for key, val in paramdf.items():
-            axsm.plot(val,
-                      c=cmap((float(key)/len(paramdf.keys()))),
-                      lw=0.25, marker='.', ms=2)
-
-        # add colorbar
-        axcb = f.add_axes([axsm.get_position().x1-.01*(nparam+1),
-                           axsm.get_position().y0,
-                           .01,
-                           axsm.get_position().y1-axsm.get_position().y0])
-
-        cbbounds = [1] + list(np.arange(2, len(paramdf.keys()) + 1, 1))
-
-
-        cb = mpl.colorbar.ColorbarBase(axcb, cmap=cmap,
-                                 orientation = 'vertical',
-                                 boundaries = [0] + cbbounds + [cbbounds[-1]+1],
-                                 spacing='proportional',
-                                 norm=mpl.colors.BoundaryNorm(cbbounds, cmap.N)
-                                 )
-        if nparam > 0: cb.set_ticks([])
-
-        smhandles += [mpl.lines.Line2D([],[], color=cmap(0))]
-        smlabels += [parameter]
-    axsm.legend(handles=smhandles, labels=smlabels, loc='upper left')
-
-    axsmbounds = list(axsm.get_position().bounds)
-    axsmbounds[2] = axsmbounds[2] - 0.015*len(interres_params)
-    axsm.set_position(axsmbounds)
-
-    for [pax, jax, [key, val]] in zip(paramaxes, jacaxes, interparams.items()):
-        if key not in interjacs: continue
-        pax.plot(*val, label=key, marker='.', ms=3, lw=0.5)
-        pax.legend(loc='upper center')
-        jax.plot(interjacs[key][0], interjacs[key][1], label=key, marker='.', ms=3, lw=0.5)
-        jax.legend(loc='upper center')
-
-    for key, val in intererrs.items():
-        if key == 'abserr':
-            axerr.semilogy(val[0],np.abs(val[1]), label=key, marker='.', ms=3, lw=0.5, c='r')
-            axerr.legend(ncol=5, loc='upper left')
-        if key == 'relerr':
-            axrelerr = axerr.twinx()
-            axrelerr.semilogy(val[0],np.abs(val[1]), label=key, marker='.', ms=3, lw=0.5, c='g')
-            axrelerr.legend(ncol=5, loc='upper right')
-
-
-
-def printresults(fit, truevals=None, startvals=False,
-                 legends=True, result_selection='all'):
-    '''
-    a function to quickly print the fit-results and the gained parameters
-
-    Parametsrs:
-    ------------
-    fit : list
-          output of performfit()-function
-    truevals : dict (default = None)
-               dictionary of the expected parameter-values (must be of the
-               same shape as the parameter-values gained from the fit).
-               if provided, the difference between the expected- and
-               fitted values is plotted
-    startvals : bool (default = False)
-                if True, the model-results using the start-values are
-                plotted as black lines
-    legends : bool (default = True)
-              indicator if legends should be plotted
-    result_selection : list-like or 'all'
-                       a list of the measurement-numbers that should be
-                       plotted (indexed starting from 0) or 'all' in case
-                       all measurements should be plotted
-    Returns:
-    ---------
-    fig : matplotlib.figure object
-    '''
-    # this is done to allow the usage of monofit-outputs as well
-    (res_lsq, R, data, inc, mask, weights,
-     res_dict, start_dict, fixed_dict) = fit.result
-
-    if result_selection == 'all':
-        result_selection = range(len(data))
-
-    # assign colors
-    colordict = {key:f'C{i%10}' for i, key in enumerate(res_dict.keys())}
-
-    # reset incidence-angles in case they have been altered beforehand
-    R.t_0 = inc
-    R.p_0 = np.zeros_like(inc)
-
-    # evaluate number of measurements
-    Nmeasurements = len(inc)
-
-    if truevals is not None:
-        truevals = {**truevals}
-
-        # generate a dictionary to assign values based on input
-        for key in truevals:
-            if np.isscalar(truevals[key]):
-                truevals[key] = np.array([truevals[key]] * Nmeasurements)
-            else:
-                truevals[key] = truevals[key]
-
-    # generate figure
-    fig = plt.figure(figsize=(14, 10))
-    ax = fig.add_subplot(211)
-    ax.set_title('Fit-results')
-
-    # plot datapoints
-    for i, j in enumerate(np.ma.masked_array(data, mask)[result_selection]):
-        ax.plot(inc[result_selection[i]], j, '.')
-
-    # reset color-cycle
-    plt.gca().set_prop_cycle(None)
-
-    # define incidence-angle range for plotting
-    incplot = np.array([np.linspace(np.min(inc), np.max(inc), 100)]
-                       * Nmeasurements)
-    # set new incidence-angles
-    R.t_0 = incplot
-    R.p_0 = np.zeros_like(incplot)
-
-    # get parameter-values
-    # for python > 3.4
-    # calc_dict = dict(**res_dict, **fixed_dict)
-    calc_dict = {**res_dict, **fixed_dict}
-
-    # calculate results
-    fitplot = fit._calc_model(R, calc_dict)
-
-    # generate a mask that hides all measurements where no data has
-    # been provided (i.e. whose parameter-results are still the startvals)
-    newmask = np.ones_like(incplot) * np.all(mask, axis=1)[:, np.newaxis]
-    fitplot = np.ma.masked_array(fitplot, newmask)
-
-    for i, val in enumerate(fitplot[result_selection]):
-        ax.plot(incplot[i], val, alpha=0.4, label=result_selection[i])
-
-    # ----------- plot start-values ------------
-    if startvals is True:
-        startplot = fit._calc_model(R, {**start_dict, **fixed_dict})
-        for i, val in enumerate(startplot[result_selection]):
-            if i == 0:
-                label = 'fitstart'
-            else:
-                label = ''
-            ax.plot(incplot[result_selection[i]], val, 'k--', linewidth=1,
-                    alpha=0.5, label=label)
-
-    if legends is True:
-        ax.legend(loc=1)
-
-    mintic = np.round(np.rad2deg(np.min(inc)) + 4.9, -1)
-    if mintic < 0.:
-        mintic = 0.
-    maxtic = np.round(np.rad2deg(np.max(inc)) + 4.9, -1)
-    if maxtic > 360.:
-        maxtic = 360.
-
-    ticks = np.arange(np.rad2deg(np.min(inc)),
-                      np.rad2deg(np.max(inc)) + 1.,
-                      (maxtic - mintic) / 10.)
-    plt.xticks(np.deg2rad(ticks), np.array(ticks, dtype=int))
-    plt.xlabel('$\\theta_0$ [deg]')
-    plt.ylabel('$I_{tot}$')
-
-    ax2 = fig.add_subplot(212)
-    ax2.set_title('Estimated parameters')
-
-
-    if truevals is not None:
-
-        # plot actual values
-        for key in truevals:
-            ax2.plot(fit.index, truevals[key],
-                     '--', alpha=0.75, color=colordict[key])
-        for key in truevals:
-            ax2.plot(fit.index, truevals[key], 'o',
-                     color=colordict[key])
-
-        param_errs = {}
-        for key in truevals:
-            param_errs[key] = res_dict[key] - truevals[key]
-
-        for key in truevals:
-            ax2.plot(fit.index, param_errs[key],
-                     ':', alpha=.25, color=colordict[key])
-        for key in truevals:
-            ax2.plot(fit.index, param_errs[key],
-                     '.', alpha=.25, color=colordict[key])
-
-        h2 = mlines.Line2D([], [], color='black', label='data',
-                           linestyle='--', alpha=0.75, marker='o')
-        h3 = mlines.Line2D([], [], color='black', label='errors',
-                           linestyle=':', alpha=0.5, marker='.')
-
-
-    # plot fitted values
-    for key in res_dict:
-        ax2.plot(fit.index,
-                 np.ma.masked_array(res_dict[key], np.all(mask, axis=1)),
-                 alpha=1., label=key, color=colordict[key])
-
-    if len(result_selection) < len(data):
-        for i, resid in enumerate(result_selection):
-            ax2.text(fit.index[resid],
-                     ax2.get_ylim()[1]*.9,
-                     resid,
-                     bbox=dict(facecolor=f'C{i}', alpha=0.5))
-
-
-
-    h1 = mlines.Line2D([], [], color='black', label='estimates',
-                       linestyle='-', alpha=0.75, marker='.')
-
-    handles, labels = ax2.get_legend_handles_labels()
-    if truevals is None:
-        plt.legend(handles=handles + [h1], loc=1)
-    else:
-        plt.legend(handles=handles + [h1, h2, h3], loc=1)
-
-    # set ticks
-    if isinstance(fit.index[0], datetime.datetime):
-        fig.autofmt_xdate()
-
-    if truevals is None:
-        plt.ylabel('Parameters')
-    else:
-        plt.ylabel('Parameters / Errors')
-
-    fig.tight_layout()
-
-    return fig
-
-
-
-def printerr(fit, newcalc=False, relative=False, result_selection='all'):
-    '''
-    a function to quickly print residuals for each measurement
-    and for each incidence-angle value
-
-    Parametsrs:
-    ------------
-    fit : list
-        output of performfit()-function
-    newcalc : bool (default = False)
-              indicator whether the residuals shall be re-calculated
-              or not.
-
-              True:
-                  the residuals are calculated using R, inc, mask,
-                  res_dict and fixed_dict from the fit-argument
-              False:
-                  the residuals are taken from the output of
-                  res_lsq from the fit-argument
-    relative : bool (default = False)
-               indicator if relative (True) or absolute (False) residuals
-               shall be plotted
     '''
 
-    (res_lsq, R, data, inc, mask, weights,
-     res_dict, start_dict, fixed_dict) = fit.result
+    def __init__(self, fit=None, **kwargs):
+        self.fit = fit
 
-    if result_selection == 'all':
-        result_selection = range(len(data))
+    def printscatter(self, fit=None, mima=None, pointsize=0.5,
+                     regression=True, newcalc=False,  **kwargs):
+        '''
+        geerate a scatterplot of modelled vs. original backscatter data
 
-    Nmeasurements = len(inc)
+        Parameters:
+        ------------
+        fit : list
+              output of monofit()-function
+        Other Parameters:
+        ------------------
+        mima : list
+               manual definition plot-boundaries via mima = [min, max]
+        pointsize : float
+                    manual specification of pointsize
+        regression : bool (default = True)
+                     indicator if the scipy.stats.linregress should be called
+                     to get the regression-line and the r^2 value
+        kwargs : -
+                 kwargs passed to matplotlib.pyplot.scatter()
+        '''
+        if fit is None:
+            fit = self.fit
 
-    if newcalc is False:
-        # get residuals from fit into desired shape for plotting
-        # Attention -> incorporate weights and mask !
-        res = np.ma.masked_array(np.reshape(res_lsq.fun, data.shape), mask)
+        (res_lsq, R, data, inc, mask, weights,
+         res_dict, start_dict, fixed_dict) = fit.result
 
-        if relative is True:
-            res = np.ma.abs(res / (res + np.ma.masked_array(data, mask)))
-        else:
-            res = np.ma.abs(res)
-    else:
-        # Alternative way of calculating the residuals
-        # (based on R, inc and res_dict)
-
+        # reset incidence-angles in case they have been altered beforehand
         R.t_0 = inc
         R.p_0 = np.zeros_like(inc)
+
+        fig = plt.figure()
+        ax = fig.add_subplot(111)
+
+        if newcalc is True:
+
+            # for python > 3.4
+            # calc_dict = dict(**res_dict, **fixed_dict)
+            calc_dict = dict((k, v) for k, v in list(res_dict.items())
+                             + list(fixed_dict.items()))
+
+            estimates = fit._calc_model(R, calc_dict)
+
+            # apply mask
+            estimates = estimates[~mask]
+            measures = data[~mask]
+
+        else:
+            # get the residuals and apply mask
+            residuals = np.reshape(res_lsq.fun, data.shape)
+            residuals = np.ma.masked_array(residuals, mask)
+            # prepare measurements
+            measures = data[~mask]
+            # calculate estimates
+            estimates = residuals[~mask] + measures
+
+        if mima is None:
+            mi = np.min((measures, estimates))
+            ma = np.max((measures, estimates))
+        else:
+            mi, ma = mima
+
+        ax.scatter(estimates, measures, s=pointsize, alpha=0.7, **kwargs)
+
+        # plot 45degree-line
+        ax.plot([mi, ma], [mi, ma], 'k--')
+
+        if fit.sig0 is True:
+            quantity = r'$\sigma_0$'
+        else:
+            quantity = 'Intensity'
+
+        if fit.dB is True:
+            scale = '[dB]'
+        else:
+            scale = ''
+
+        ax.set_xlabel('modelled ' + quantity + scale)
+        ax.set_ylabel('measured ' + quantity + scale)
+
+        if regression is True:
+            from scipy.stats import linregress
+            # evaluate linear regression to get r-value etc.
+            slope, intercept, r_value, p_value, std_err = linregress(estimates,
+                                                                     measures)
+
+            ax.plot(np.sort(measures),
+                    intercept + slope * np.sort(measures), 'r--', alpha=0.4)
+
+            ax.text(0.8, .1, '$R^2$ = ' + str(np.round(r_value**2, 2)),
+                    horizontalalignment='center',
+                    verticalalignment='center',
+                    transform=ax.transAxes)
+
+        return fig
+
+
+    def printresults(self, fit=None, truevals=None, startvals=False,
+                 legends=True, result_selection='all'):
+        '''
+        a function to quickly print the fit-results and the gained parameters
+
+        Parametsrs:
+        ------------
+        fit : list
+              output of performfit()-function
+        truevals : dict (default = None)
+                   dictionary of the expected parameter-values (must be of the
+                   same shape as the parameter-values gained from the fit).
+                   if provided, the difference between the expected- and
+                   fitted values is plotted
+        startvals : bool (default = False)
+                    if True, the model-results using the start-values are
+                    plotted as black lines
+        legends : bool (default = True)
+                  indicator if legends should be plotted
+        result_selection : list-like or 'all'
+                           a list of the measurement-numbers that should be
+                           plotted (indexed starting from 0) or 'all' in case
+                           all measurements should be plotted
+        Returns:
+        ---------
+        fig : matplotlib.figure object
+        '''
+
+        if fit is None:
+            fit = self.fit
+
+        # this is done to allow the usage of monofit-outputs as well
+        (res_lsq, R, data, inc, mask, weights,
+         res_dict, start_dict, fixed_dict) = fit.result
+
+        if result_selection == 'all':
+            result_selection = range(len(data))
+
+        # assign colors
+        colordict = {key:f'C{i%10}' for i, key in enumerate(res_dict.keys())}
+
+        # reset incidence-angles in case they have been altered beforehand
+        R.t_0 = inc
+        R.p_0 = np.zeros_like(inc)
+
+        # evaluate number of measurements
+        Nmeasurements = len(inc)
+
+        if truevals is not None:
+            truevals = {**truevals}
+
+            # generate a dictionary to assign values based on input
+            for key in truevals:
+                if np.isscalar(truevals[key]):
+                    truevals[key] = np.array([truevals[key]] * Nmeasurements)
+                else:
+                    truevals[key] = truevals[key]
+
+        # generate figure
+        fig = plt.figure(figsize=(14, 10))
+        ax = fig.add_subplot(211)
+        ax.set_title('Fit-results')
+
+        # plot datapoints
+        for i, j in enumerate(np.ma.masked_array(data, mask)[result_selection]):
+            ax.plot(inc[result_selection[i]], j, '.')
+
+        # reset color-cycle
+        plt.gca().set_prop_cycle(None)
+
+        # define incidence-angle range for plotting
+        incplot = np.array([np.linspace(np.min(inc), np.max(inc), 100)]
+                           * Nmeasurements)
+        # set new incidence-angles
+        R.t_0 = incplot
+        R.p_0 = np.zeros_like(incplot)
+
+        # get parameter-values
+        # for python > 3.4
+        # calc_dict = dict(**res_dict, **fixed_dict)
+        calc_dict = {**res_dict, **fixed_dict}
+
+        # calculate results
+        fitplot = fit._calc_model(R, calc_dict)
+
+        # generate a mask that hides all measurements where no data has
+        # been provided (i.e. whose parameter-results are still the startvals)
+        newmask = np.ones_like(incplot) * np.all(mask, axis=1)[:, np.newaxis]
+        fitplot = np.ma.masked_array(fitplot, newmask)
+
+        for i, val in enumerate(fitplot[result_selection]):
+            ax.plot(incplot[i], val, alpha=0.4, label=result_selection[i])
+
+        # ----------- plot start-values ------------
+        if startvals is True:
+            startplot = fit._calc_model(R, {**start_dict, **fixed_dict})
+            for i, val in enumerate(startplot[result_selection]):
+                if i == 0:
+                    label = 'fitstart'
+                else:
+                    label = ''
+                ax.plot(incplot[result_selection[i]], val, 'k--', linewidth=1,
+                        alpha=0.5, label=label)
+
+        if legends is True:
+            ax.legend(loc=1)
+
+        mintic = np.round(np.rad2deg(np.min(inc)) + 4.9, -1)
+        if mintic < 0.:
+            mintic = 0.
+        maxtic = np.round(np.rad2deg(np.max(inc)) + 4.9, -1)
+        if maxtic > 360.:
+            maxtic = 360.
+
+        ticks = np.arange(np.rad2deg(np.min(inc)),
+                          np.rad2deg(np.max(inc)) + 1.,
+                          (maxtic - mintic) / 10.)
+        plt.xticks(np.deg2rad(ticks), np.array(ticks, dtype=int))
+        plt.xlabel('$\\theta_0$ [deg]')
+        plt.ylabel('$I_{tot}$')
+
+        ax2 = fig.add_subplot(212)
+        ax2.set_title('Estimated parameters')
+
+
+        if truevals is not None:
+
+            # plot actual values
+            for key in truevals:
+                ax2.plot(fit.index, truevals[key],
+                         '--', alpha=0.75, color=colordict[key])
+            for key in truevals:
+                ax2.plot(fit.index, truevals[key], 'o',
+                         color=colordict[key])
+
+            param_errs = {}
+            for key in truevals:
+                param_errs[key] = res_dict[key] - truevals[key]
+
+            for key in truevals:
+                ax2.plot(fit.index, param_errs[key],
+                         ':', alpha=.25, color=colordict[key])
+            for key in truevals:
+                ax2.plot(fit.index, param_errs[key],
+                         '.', alpha=.25, color=colordict[key])
+
+            h2 = mlines.Line2D([], [], color='black', label='data',
+                               linestyle='--', alpha=0.75, marker='o')
+            h3 = mlines.Line2D([], [], color='black', label='errors',
+                               linestyle=':', alpha=0.5, marker='.')
+
+
+        # plot fitted values
+        for key in res_dict:
+            ax2.plot(fit.index,
+                     np.ma.masked_array(res_dict[key], np.all(mask, axis=1)),
+                     alpha=1., label=key, color=colordict[key])
+
+        if len(result_selection) < len(data):
+            for i, resid in enumerate(result_selection):
+                ax2.text(fit.index[resid],
+                         ax2.get_ylim()[1]*.9,
+                         resid,
+                         bbox=dict(facecolor=f'C{i}', alpha=0.5))
+
+
+
+        h1 = mlines.Line2D([], [], color='black', label='estimates',
+                           linestyle='-', alpha=0.75, marker='.')
+
+        handles, labels = ax2.get_legend_handles_labels()
+        if truevals is None:
+            plt.legend(handles=handles + [h1], loc=1)
+        else:
+            plt.legend(handles=handles + [h1, h2, h3], loc=1)
+
+        # set ticks
+        if isinstance(fit.index[0], datetime.datetime):
+            fig.autofmt_xdate()
+
+        if truevals is None:
+            plt.ylabel('Parameters')
+        else:
+            plt.ylabel('Parameters / Errors')
+
+        fig.tight_layout()
+
+        return fig
+
+
+    def printerr(self, fit=None, newcalc=False, relative=False,
+                 result_selection='all'):
+        '''
+        a function to quickly print residuals for each measurement
+        and for each incidence-angle value
+
+        Parametsrs:
+        ------------
+        fit : list
+            output of performfit()-function
+        newcalc : bool (default = False)
+                  indicator whether the residuals shall be re-calculated
+                  or not.
+
+                  True:
+                      the residuals are calculated using R, inc, mask,
+                      res_dict and fixed_dict from the fit-argument
+                  False:
+                      the residuals are taken from the output of
+                      res_lsq from the fit-argument
+        relative : bool (default = False)
+                   indicator if relative (True) or absolute (False) residuals
+                   shall be plotted
+        '''
+
+        if fit is None:
+            fit = self.fit
+
+        (res_lsq, R, data, inc, mask, weights,
+         res_dict, start_dict, fixed_dict) = fit.result
+
+        if result_selection == 'all':
+            result_selection = range(len(data))
+
+        if newcalc is False:
+            # get residuals from fit into desired shape for plotting
+            # Attention -> incorporate weights and mask !
+            res = np.ma.masked_array(np.reshape(res_lsq.fun, data.shape), mask)
+
+            if relative is True:
+                res = np.ma.abs(res / (res + np.ma.masked_array(data, mask)))
+            else:
+                res = np.ma.abs(res)
+        else:
+            # Alternative way of calculating the residuals
+            # (based on R, inc and res_dict)
+
+            R.t_0 = inc
+            R.p_0 = np.zeros_like(inc)
+
+            # for python > 3.4
+            # calc_dict = dict(**res_dict, **fixed_dict)
+            calc_dict = dict((k, v) for k, v in list(res_dict.items())
+                             + list(fixed_dict.items()))
+
+            estimates = fit._calc_model(R, calc_dict)
+            # calculate the residuals based on masked arrays
+            masked_estimates = np.ma.masked_array(estimates, mask=mask)
+            masked_data = np.ma.masked_array(data, mask=mask)
+
+            res = np.ma.sqrt((masked_estimates - masked_data)**2)
+
+            if relative is True:
+                res = res / masked_estimates
+
+        # apply mask to data and incidence-angles (and convert to degree)
+        inc = np.ma.masked_array(np.rad2deg(inc), mask=mask)
+        data = np.ma.masked_array(data, mask=mask)
+
+        # make new figure
+        fig = plt.figure(figsize=(14, 10))
+        ax = fig.add_subplot(212)
+        if relative is True:
+            ax.set_title('Mean relative residual per measurement')
+        else:
+            ax.set_title('Mean absolute residual per measurement')
+
+        ax2 = fig.add_subplot(211)
+        if relative is True:
+            ax2.set_title('Relative residuals per incidence-angle')
+        else:
+            ax2.set_title('Residuals per incidence-angle')
+
+        # the use of masked arrays might cause python 2 compatibility issues!
+        ax.plot(fit.index[result_selection], res[result_selection], '.', alpha=0.5)
+
+        # plot mean residual for each measurement
+        ax.plot(fit.index[result_selection], np.ma.mean(res[result_selection], axis=1),
+                   'k', linewidth=3, marker='o', fillstyle='none')
+
+        # plot total mean of mean residuals per measurement
+        ax.plot(fit.index[result_selection],
+                   [np.ma.mean(np.ma.mean(res[result_selection], axis=1))] * len(result_selection),
+                   'k--')
+
+        # add some legends
+        res_h = mlines.Line2D(
+            [], [], color='black', label='Mean res.  per measurement',
+            linestyle='-', linewidth=3, marker='o', fillstyle='none')
+        res_h_dash = mlines.Line2D(
+            [], [], color='black', linestyle='--', label='Average mean res.',
+            linewidth=1, fillstyle='none')
+
+        res_h_dots = mlines.Line2D(
+            [], [], color='black', label='Residuals',
+            linestyle='-', linewidth=0, marker='.', alpha=0.5)
+
+        handles, labels = ax.get_legend_handles_labels()
+        ax.legend(handles=handles + [res_h_dots] + [res_h] + [res_h_dash],
+                     loc=1)
+
+        ax.set_ylabel('Residual')
+
+    #        # evaluate mean residuals per incidence-angle
+        meanincs = np.ma.unique(np.concatenate(inc[result_selection]))
+        mean = np.full_like(meanincs, 0.)
+
+        for a, incval in enumerate(meanincs):
+            select = np.where(inc[result_selection] == incval)
+            res_selected = res[result_selection][select[0][:, np.newaxis],
+                               select[1][:, np.newaxis]]
+            mean[a] = np.ma.mean(res_selected)
+
+        sortpattern = np.argsort(meanincs)
+        meanincs = meanincs[sortpattern]
+        mean = mean[sortpattern]
+
+        # plot residuals per incidence-angle for each measurement
+        for i, resval in enumerate(res[result_selection]):
+            sortpattern = np.argsort(inc[result_selection[i]])
+            ax2.plot(inc[result_selection[i]][sortpattern], resval[sortpattern],
+                        ':', alpha=0.5, marker='.')
+
+        # plot mean residual per incidence-angle
+        ax2.plot(meanincs, mean,
+                    'k', linewidth=3, marker='o', fillstyle='none')
+
+        # add some legends
+        res_h2 = mlines.Line2D(
+            [], [], color='black', label='Mean res.  per inc-angle',
+            linestyle='-', linewidth=3, marker='o', fillstyle='none')
+        res_h_lines = mlines.Line2D(
+            [], [], color='black', label='Residuals',
+            linestyle=':', alpha=0.5)
+
+        handles2, labels2 = ax2.get_legend_handles_labels()
+        ax2.legend(handles=handles2 + [res_h_lines] + [res_h2], loc=1)
+
+        ax2.set_xlabel('$\\theta_0$ [deg]')
+        ax2.set_ylabel('Residual')
+
+        # find minimum and maximum incidence angle
+        maxinc = np.max(inc)
+        mininc = np.min(inc)
+
+        ax2.set_xlim(np.floor(mininc) - 1,
+                        np.ceil(maxinc) + 1)
+
+        # set major and minor ticks
+        ax2.xaxis.set_major_locator(plt.MultipleLocator(1))
+        ax2.xaxis.set_major_formatter(plt.FormatStrFormatter('%d'))
+        ax2.xaxis.set_minor_locator(plt.MultipleLocator(.25))
+
+        # set ticks
+        if isinstance(fit.index[0], datetime.datetime):
+            plt.setp(ax.get_xticklabels(), rotation=30, ha='right')
+
+        fig.tight_layout()
+
+        return fig
+
+
+    def plot_interres(self, fit=None, params = None,
+                      cmaps=None):
+        '''
+        a function to plot the intermediate-results
+        (the data is only available if rtfits.monofit has been called with
+        the argument intermediate_results=True!)
+
+        Parameters:
+        -------------
+        fit : rtfits object
+              the rtfits-object containing the fit-results
+        params : list
+            a list of parameter-names that are intended to be plotted
+            as timeseries.
+        cmaps : list
+            a list of the colormaps used to plot the parameter variations
+        '''
+
+        if fit is None:
+            fit = self.fit
+
+        try:
+            fit.intermediate_results
+        except AttributeError:
+            assert False, ('No intermediate results are found, you must run' +
+                           ' monofit() with intermediate_results=True flag!')
+
+        if params is None:
+            params = fit.result[6].keys()
+
+
+        if cmaps is None:
+            cmaps = ['Reds', 'Greens', 'Blues', 'Purples', 'Oranges', 'Greys',
+                     'YlOrBr', 'YlOrRd', 'OrRd', 'PuRd', 'RdPu', 'BuPu',
+                     'GnBu', 'PuBu', 'YlGnBu', 'PuBuGn', 'BuGn', 'YlGn']
+
+        interparams = {}
+        for i, valdict in enumerate(fit.intermediate_results['parameters']):
+            for key, val in valdict.items():
+                if key in interparams:
+                    interparams[key] += [[i, np.mean(val)]]
+                else:
+                    interparams[key] = [[i, np.mean(val)]]
+        intererrs = {}
+        for i, valdict in enumerate(fit.intermediate_results['residuals']):
+            for key, val in valdict.items():
+                if key in intererrs:
+                    intererrs[key] += [[i, np.mean(val)]]
+                else:
+                    intererrs[key] = [[i, np.mean(val)]]
+        interjacs = {}
+        for i, valdict in enumerate(fit.intermediate_results['jacobian']):
+            for key, val in valdict.items():
+                if key in interjacs:
+                    interjacs[key] += [[i, np.mean(val)]]
+                else:
+                    interjacs[key] = [[i, np.mean(val)]]
+
+        for i in [interparams, intererrs, interjacs]:
+            for key, val in i.items():
+                i[key] = np.array(val).T
+
+        interres_params = {}
+        for key in params:
+            interres_p = pd.concat([pd.DataFrame(valdict[key],
+                                                 fit.index.drop_duplicates(),
+                                    columns=[i])
+                                    for i, valdict in enumerate(
+                                            fit.intermediate_results['parameters'])
+                                    ], axis=1)
+            interres_params[key] = interres_p
+
+
+        f = plt.figure(figsize=(15,10))
+        f.subplots_adjust(top=0.98, left=0.05, right=0.95)
+        gs = mpl.gridspec.GridSpec(4, len(interjacs),
+                               #width_ratios=[1, 2],
+                               #height_ratios=[1, 2, 1]
+                               )
+        axsm = plt.subplot(gs[0,:])
+        axerr = plt.subplot(gs[1,:])
+        paramaxes, jacaxes = [], []
+        for i in range(len(interjacs)):
+            paramaxes += [plt.subplot(gs[2,i])]
+            jacaxes += [plt.subplot(gs[3,i])]
+
+        smhandles, smlabels = [],[]
+        for nparam, [parameter, paramdf] in enumerate(interres_params.items()):
+
+            cmap = plt.get_cmap(cmaps[nparam])
+
+            for key, val in paramdf.items():
+                axsm.plot(val,
+                          c=cmap((float(key)/len(paramdf.keys()))),
+                          lw=0.25, marker='.', ms=2)
+
+            # add colorbar
+            axcb = f.add_axes([axsm.get_position().x1-.01*(nparam+1),
+                               axsm.get_position().y0,
+                               .01,
+                               axsm.get_position().y1-axsm.get_position().y0])
+
+            cbbounds = [1] + list(np.arange(2, len(paramdf.keys()) + 1, 1))
+
+
+            cb = mpl.colorbar.ColorbarBase(axcb, cmap=cmap,
+                                     orientation = 'vertical',
+                                     boundaries = [0] + cbbounds + [cbbounds[-1]+1],
+                                     spacing='proportional',
+                                     norm=mpl.colors.BoundaryNorm(cbbounds, cmap.N)
+                                     )
+            if nparam > 0: cb.set_ticks([])
+
+            smhandles += [mpl.lines.Line2D([],[], color=cmap(.9))]
+            smlabels += [parameter]
+        axsm.legend(handles=smhandles, labels=smlabels, loc='upper left')
+
+        axsmbounds = list(axsm.get_position().bounds)
+        axsmbounds[2] = axsmbounds[2] - 0.015*len(interres_params)
+        axsm.set_position(axsmbounds)
+
+        for [pax, jax, [key, val]] in zip(paramaxes, jacaxes, interparams.items()):
+            if key not in interjacs: continue
+            pax.plot(*val, label=key, marker='.', ms=3, lw=0.5)
+            pax.legend(loc='upper center')
+            jax.plot(interjacs[key][0], interjacs[key][1], label=key, marker='.', ms=3, lw=0.5)
+            jax.legend(loc='upper center')
+
+        for key, val in intererrs.items():
+            if key == 'abserr':
+                axerr.semilogy(val[0],np.abs(val[1]), label=key, marker='.', ms=3, lw=0.5, c='r')
+                axerr.legend(ncol=5, loc='upper left')
+            if key == 'relerr':
+                axrelerr = axerr.twinx()
+                axrelerr.semilogy(val[0],np.abs(val[1]), label=key, marker='.', ms=3, lw=0.5, c='g')
+                axrelerr.legend(ncol=5, loc='upper right')
+
+
+        return f
+
+
+    def printsig0timeseries(self, fit=None, dB=True, sig0=True, params=None,
+                            printtot = True, printsurf = True,
+                            printvol = True, printint = True,
+                            printorig = True, months = None, years = None,
+                            ylim=None, printinc = True):
+        '''
+        Print individual contributions, resulting parameters and the
+        reference dataset of an rt1.rtfits object as timeseries.
+
+        Parameters:
+        -------------
+        fit : rtfits object
+              the rtfits-object containing the fit-results
+        dB : bool (default = True)
+             indicator if the plot is intended to be in dB or linear units
+        sig0 : bool (default = True)
+             indicator if the plot should display sigma_0 (sig0) or intensity (I)
+             The applied relation is: sig0 = 4.*pi*cos(theta) * I
+        params: list
+                a list of parameter-names that should be overprinted
+                (the names must coincide with the arguments of set_V_SRF())
+        printtot, printsurf, printvol, printint, printorig : bool
+                indicators if the corresponding components should be plotted
+        months : list of int (default = None)
+                 a list of months to plot (if None, all will be plotted)
+        years : list of int (default = None)
+                a list of years to select (if None, all will be plotted)
+        ylim : tuple
+               a tuple of (ymin, ymax) that will be used as boundaries for the
+               y-axis
+        printinc : bool (default = True)
+                   indicator if the incidence-angle dependency should be plotted
+                   (in a separate plot alongside the timeseries)
+        '''
+
+        if fit is None:
+            fit = self.fit
+
+        # get mask
+        (_, _, data, _, mask, _, _, _, _) = fit.result
+        # get incidence-angles
+        inc_array = np.ma.masked_array(fit.result[1].t_0, mask)
+        inc = inc_array.compressed()
+        # get input dataset
+        data = np.ma.masked_array(data, mask)
+
+        def dBsig0convert(val):
+            # if results are provided in dB convert them to linear units
+            if fit.dB is True: val = 10**(val/10.)
+            # convert sig0 to intensity
+            if sig0 is False and fit.sig0 is True:
+                val = val/(4.*np.pi*np.cos(inc))
+            # convert intensity to sig0
+            if sig0 is True and fit.sig0 is False:
+                val = 4.*np.pi*np.cos(inc)*val
+            # if dB output is required, convert to dB
+            if dB is True: val = 10.*np.log10(val)
+            return val
+
+        res_dict = {}
+        # add fitted parameters
+        res_dict.update(fit.result[6])
+        # add constant values
+        res_dict.update(fit.result[-1])
+
+        # calculate individual contributions
+        contrib_array = fit._calc_model(R=fit.result[1],
+                                        res_dict=res_dict,
+                                        return_components=True)
+
+        # apply mask and convert to pandas dataframe
+        contrib_array = [np.ma.masked_array(con, mask) for con in contrib_array]
+        contrib_array += [data, inc_array]
+
+        contrib = []
+        for i, cont in enumerate(contrib_array):
+            contrib += [pd.concat([pd.DataFrame(i, index = fit.index) for i in cont.T])[0]]
+
+        contrib = pd.concat(contrib,
+                            keys=['tot', 'surf', 'vol', 'inter',
+                                  '$\\sigma_0$ dataset', 'inc'], axis=1).dropna()
+
+        # convert units
+        contrib[['tot', 'surf', 'vol', 'inter',
+                 '$\\sigma_0$ dataset']] = contrib[[
+                         'tot', 'surf', 'vol', 'inter', '$\\sigma_0$ dataset'
+                         ]].apply(dBsig0convert)
+
+        # drop unneeded columns
+        if fit.result[1].int_Q is False or printint is False:
+            contrib = contrib.drop('inter', axis=1)
+        if printtot is False: contrib = contrib.drop('tot', axis=1)
+        if printsurf is False: contrib = contrib.drop('surf', axis=1)
+        if printvol is False: contrib = contrib.drop('vol', axis=1)
+        if printorig is False:
+            contrib = contrib.drop('$\\sigma_0$ dataset', axis=1)
+
+        # select years and months
+        if years is not None:
+            contrib = contrib.loc[contrib.index.year.isin(years)]
+        if months is not None:
+            contrib = contrib.loc[contrib.index.month.isin(months)]
+
+        # print incidence-angle dependency
+        if printinc is True:
+            f, [ax, ax_inc] = plt.subplots(ncols=2, figsize=(15,5),
+                                           gridspec_kw={'width_ratios':[3,1]},
+                                           sharey=True)
+            f.subplots_adjust(left=0.05, right=0.98, top=0.98,
+                              bottom=0.1, wspace=0.1)
+
+            # -------------------
+            color = {'tot':'r', 'surf':'b', 'vol':'g', 'inter':'y',
+                     '$\\sigma_0$ dataset':'k'}
+
+            groupedcontrib = contrib.groupby(contrib.index)
+
+            #return contrib, groupedcontrib
+            for label in contrib.keys():
+                if label in ['inc']: continue
+                a=np.rad2deg(rectangularize([x.values for _, x in groupedcontrib['inc']])).T
+                b=np.array(rectangularize([x.values for _, x in groupedcontrib[label]])).T
+                x = (np.array([a,b]).T)
+
+                l_col = mpl.collections.LineCollection(x,linewidth =.25, label='x',
+                                          color=color[label], alpha = 0.5)
+                ax_inc.add_collection(l_col)
+                ax_inc.scatter(a, b, color=color[label], s=1)
+                ax_inc.set_xlim(a.min(), a.max())
+                ax_inc.set_xlabel('$\\theta_0$')
+            ax_inc.set_xlabel('$\\theta_0$')
+
+
+        else:
+            f, ax = plt.subplots(figsize=(12,5))
+            f.subplots_adjust(left=0.05, right=0.98, top=0.98,
+                              bottom=0.1, wspace=0.05)
+
+        for label, val in contrib.items():
+            if label in ['inc']: continue
+            color = {'tot':'r', 'surf':'b', 'vol':'g', 'inter':'y'}
+            if printorig is True: color['$\\sigma_0$ dataset'] = 'k'
+            ax.plot(val.sort_index(), linewidth =.25, marker='.',
+                    ms=2, label=label, color=color[label], alpha = 0.5)
+        # overprint parameters
+        if params != None:
+            paramdf_dict = {}
+            # add fitted parameters
+            paramdf_dict.update(fit.result[6])
+            # add constant values
+            paramdf_dict.update(fit.result[-1])
+
+            paramdf = pd.DataFrame(paramdf_dict,
+                                   index = fit.index).sort_index()
+            if years is not None:
+                paramdf = paramdf.loc[paramdf.index.year.isin(years)]
+            if months is not None:
+                paramdf = paramdf.loc[paramdf.index.month.isin(months)]
+
+            pax = ax.twinx()
+            for k in params:
+                pax.plot(paramdf[k], lw=1, marker='.', ms=2, label=k)
+            pax.legend(loc='upper right', ncol=5)
+            pax.set_ylabel('parameter-values')
+
+        # format datetime index
+        ax.xaxis.set_minor_locator(mpl.dates.MonthLocator())
+        ax.xaxis.set_minor_formatter(mpl.dates.DateFormatter('%m'))
+        ax.xaxis.set_major_locator(mpl.dates.YearLocator())
+        ax.xaxis.set_major_formatter(mpl.dates.DateFormatter('\n%Y'))
+
+        # set ylabels
+        if sig0 is True:
+            label = '$\\sigma_0$'
+        else:
+            label = 'Intensity'
+        if dB is True: label += ' [dB]'
+        ax.set_ylabel(label)
+
+        # generate legend
+        hand, lab = ax.get_legend_handles_labels()
+        lab, unique_ind = np.unique(lab, return_index=True)
+        ax.legend(handles = list(np.array(hand)[unique_ind]),
+                  labels=list(lab), loc='upper left',
+                  ncol=5)
+
+        if ylim is not None:
+            ax.set_ylim(ylim)
+
+        return f
+
+    def printsingle(self, fit=None, fit_numbers=None, fit_indexes=None,
+                    hexbinQ=True, hexbinargs={},
+                    convertTodB=False):
+        '''
+        a function to investigate the quality of the individual fits
+
+
+
+        Parameters:
+        ------------
+        fit : list
+              output of the monofit()-function
+        fit_numbers : list
+                      a list containing the position of the measurements
+                      that should be plotted (starting from 0)
+        fit_indexes : list
+                      a list containing the index-values of the measurements
+                      that should be plotted
+
+        Other Parameters:
+        ------------------
+        hexbinQ : bool (default = False)
+                  indicator if a hexbin-plot should be underlayed
+                  that shows the distribution of the datapoints
+        hexbinargs : dict
+                     a dict containing arguments to customize the hexbin-plot
+        convertTodB : bool (default=False)
+                      if set to true, the datasets will be converted to dB
+        '''
+        if fit is None:
+            fit = self.fit
+
+        if fit_numbers is not None and fit_indexes is not None:
+            assert False, 'please provide EITHER fit_numbers OR fit_indexes!'
+        elif fit_indexes is not None:
+            fit_numbers = np.where(fit.index.isin(fit_indexes))[0]
+        elif fit_numbers is None and fit_indexes is None:
+            fit_numbers = range(len(fit.index))
+
+        # function to generate colormap that fades between colors
+        def CustomCmap(from_rgb, to_rgb):
+
+            # from color r,g,b
+            r1, g1, b1 = from_rgb
+
+            # to color r,g,b
+            r2, g2, b2 = to_rgb
+
+            cdict = {'red': ((0, r1, r1),
+                             (1, r2, r2)),
+                     'green': ((0, g1, g1),
+                               (1, g2, g2)),
+                     'blue': ((0, b1, b1),
+                              (1, b2, b2))}
+
+            cmap = LinearSegmentedColormap('custom_cmap', cdict)
+            return cmap
+
+        (res_lsq, R, data, inc, mask, weights,
+         res_dict, start_dict, fixed_dict) = fit.result
 
         # for python > 3.4
         # calc_dict = dict(**res_dict, **fixed_dict)
@@ -1450,117 +1541,70 @@ def printerr(fit, newcalc=False, relative=False, result_selection='all'):
                          + list(fixed_dict.items()))
 
         estimates = fit._calc_model(R, calc_dict)
-        # calculate the residuals based on masked arrays
-        masked_estimates = np.ma.masked_array(estimates, mask=mask)
-        masked_data = np.ma.masked_array(data, mask=mask)
 
-        res = np.ma.sqrt((masked_estimates - masked_data)**2)
 
-        if relative is True:
-            res = res / masked_estimates
+        fig = plt.figure()
+        ax = fig.add_subplot(111)
 
-    # apply mask to data and incidence-angles (and convert to degree)
-    inc = np.ma.masked_array(np.rad2deg(inc), mask=mask)
-    data = np.ma.masked_array(data, mask=mask)
+        for m_i, m in enumerate(fit_numbers):
 
-    # make new figure
-    figres = plt.figure(figsize=(14, 10))
-    axres = figres.add_subplot(212)
-    if relative is True:
-        axres.set_title('Mean relative residual per measurement')
-    else:
-        axres.set_title('Mean absolute residual per measurement')
+            if convertTodB is True:
+                y = 10.*np.log10(estimates[m][~mask[m]])
+            else:
+                y = estimates[m][~mask[m]]
 
-    axres2 = figres.add_subplot(211)
-    if relative is True:
-        axres2.set_title('Relative residuals per incidence-angle')
-    else:
-        axres2.set_title('Residuals per incidence-angle')
+            # plot data
+            label = fit.index[m]
 
-    # the use of masked arrays might cause python 2 compatibility issues!
-    axres.plot(fit.index[result_selection], res[result_selection], '.', alpha=0.5)
+            xdata = np.rad2deg(inc[m][~mask[m]])
 
-    # plot mean residual for each measurement
-    axres.plot(fit.index[result_selection], np.ma.mean(res[result_selection], axis=1),
-               'k', linewidth=3, marker='o', fillstyle='none')
+            if convertTodB is True:
+                ydata = 10.*np.log10(data[m][~mask[m]])
+            else:
+                ydata = data[m][~mask[m]]
 
-    # plot total mean of mean residuals per measurement
-    axres.plot(fit.index[result_selection],
-               [np.ma.mean(np.ma.mean(res[result_selection], axis=1))] * len(result_selection),
-               'k--')
+            # get color that will be applied to the next line drawn
+            dummy, = ax.plot(xdata[0], ydata[0], '.', alpha=0.)
+            color = dummy.get_color()
 
-    # add some legends
-    res_h = mlines.Line2D(
-        [], [], color='black', label='Mean res.  per measurement',
-        linestyle='-', linewidth=3, marker='o', fillstyle='none')
-    res_h_dash = mlines.Line2D(
-        [], [], color='black', linestyle='--', label='Average mean res.',
-        linewidth=1, fillstyle='none')
+            if hexbinQ is True:
+                args = dict(gridsize=15, mincnt=1,
+                            linewidths=0., vmin=0.5, alpha=0.7)
+                args.update(hexbinargs)
 
-    res_h_dots = mlines.Line2D(
-        [], [], color='black', label='Residuals',
-        linestyle='-', linewidth=0, marker='.', alpha=0.5)
+                # evaluate the hexbinplot once to get the maximum number of
+                # datapoints within a single hexagonal (used for normalization)
+                dummyargs = args.copy()
+                dummyargs.update({'alpha': 0.})
+                hb = ax.hexbin(xdata, ydata, **dummyargs)
 
-    handles, labels = axres.get_legend_handles_labels()
-    axres.legend(handles=handles + [res_h_dots] + [res_h] + [res_h_dash],
-                 loc=1)
+                # generate colormap that fades from white to the color
+                # of the plotted data  (asdf.get_color())
+                cmap = CustomCmap([1.00, 1.00, 1.00],
+                                  plt.cm.colors.hex2color(color))
+                # setup correct normalizing instance
+                norm = Normalize(vmin=0, vmax=hb.get_array().max())
 
-    axres.set_ylabel('Residual')
+                ax.hexbin(xdata, ydata, cmap=cmap, norm=norm, **args)
 
-#        # evaluate mean residuals per incidence-angle
-    meanincs = np.ma.unique(np.concatenate(inc[result_selection]))
-    mean = np.full_like(meanincs, 0.)
+            # plot datapoints
+            asdf, = ax.plot(xdata, ydata, '.',
+                            color=color, alpha=1.,
+                            label=label, markersize=10)
 
-    for a, incval in enumerate(meanincs):
-        select = np.where(inc[result_selection] == incval)
-        res_selected = res[result_selection][select[0][:, np.newaxis],
-                           select[1][:, np.newaxis]]
-        mean[a] = np.ma.mean(res_selected)
+            # plot results
+            iii = inc[m][~mask[m]]
+            ax.plot(np.rad2deg(iii[np.argsort(iii)]), y[np.argsort(iii)],
+                    '-', color='w', linewidth=3)
 
-    sortpattern = np.argsort(meanincs)
-    meanincs = meanincs[sortpattern]
-    mean = mean[sortpattern]
+            ax.plot(np.rad2deg(iii[np.argsort(iii)]), y[np.argsort(iii)],
+                    '-', color=asdf.get_color(), linewidth=2)
 
-    # plot residuals per incidence-angle for each measurement
-    for i, resval in enumerate(res[result_selection]):
-        sortpattern = np.argsort(inc[result_selection[i]])
-        axres2.plot(inc[result_selection[i]][sortpattern], resval[sortpattern],
-                    ':', alpha=0.5, marker='.')
+        ax.set_xlabel('$\\theta_0$ [deg]')
+        ax.set_ylabel('$\\sigma_0$ [dB]')
 
-    # plot mean residual per incidence-angle
-    axres2.plot(meanincs, mean,
-                'k', linewidth=3, marker='o', fillstyle='none')
 
-    # add some legends
-    res_h2 = mlines.Line2D(
-        [], [], color='black', label='Mean res.  per inc-angle',
-        linestyle='-', linewidth=3, marker='o', fillstyle='none')
-    res_h_lines = mlines.Line2D(
-        [], [], color='black', label='Residuals',
-        linestyle=':', alpha=0.5)
+        ax.legend(title='# Measurement')
 
-    handles2, labels2 = axres2.get_legend_handles_labels()
-    axres2.legend(handles=handles2 + [res_h_lines] + [res_h2], loc=1)
+        return fig
 
-    axres2.set_xlabel('$\\theta_0$ [deg]')
-    axres2.set_ylabel('Residual')
-
-    # find minimum and maximum incidence angle
-    maxinc = np.max(inc)
-    mininc = np.min(inc)
-
-    axres2.set_xlim(np.floor(mininc) - 1,
-                    np.ceil(maxinc) + 1)
-
-    # set major and minor ticks
-    axres2.xaxis.set_major_locator(plt.MultipleLocator(1))
-    axres2.xaxis.set_major_formatter(plt.FormatStrFormatter('%d'))
-    axres2.xaxis.set_minor_locator(plt.MultipleLocator(.25))
-
-    # set ticks
-    if isinstance(fit.index[0], datetime.datetime):
-        plt.setp(axres.get_xticklabels(), rotation=30, ha='right')
-
-    figres.tight_layout()
-
-    return figres
